feat: add PrimeStatusBanner to Account page
