+++ src/pages/Account/AccountBreakdown/Summary/index.tsx
@@ -9,6 +9,7 @@ import { formatCentsToReadableValue, formatPercentageToReadableValue } from 'uti
 
 import { SAFE_BORROW_LIMIT_PERCENTAGE } from 'constants/safeBorrowLimitPercentage';
 
+import Section from '../Section';
 import { useStyles } from './styles';
 import TEST_IDS from './testIds';
 import useExtractData from './useExtractData';
@@ -53,79 +54,81 @@ export const Summary: React.FC<SummaryProps> = ({
 
   const cells: Cell[] = [
     {
-      label: t('account.marketBreakdown.cellGroup.netApy'),
+      label: t('account.summary.cellGroup.netApy'),
       value: formatPercentageToReadableValue(netApyPercentage),
       tooltip: displayTotalVaultStake
-        ? t('account.marketBreakdown.cellGroup.netApyWithVaultStakeTooltip')
-        : t('account.marketBreakdown.cellGroup.netApyTooltip'),
+        ? t('account.summary.cellGroup.netApyWithVaultStakeTooltip')
+        : t('account.summary.cellGroup.netApyTooltip'),
       color: styles.getNetApyColor({ netApyPercentage: netApyPercentage || 0 }),
     },
     {
-      label: t('account.marketBreakdown.cellGroup.dailyEarnings'),
+      label: t('account.summary.cellGroup.dailyEarnings'),
       value: formatCentsToReadableValue({ value: dailyEarningsCents }),
     },
     {
-      label: t('account.marketBreakdown.cellGroup.totalSupply'),
+      label: t('account.summary.cellGroup.totalSupply'),
       value: formatCentsToReadableValue({ value: totalSupplyCents }),
     },
     {
-      label: t('account.marketBreakdown.cellGroup.totalBorrow'),
+      label: t('account.summary.cellGroup.totalBorrow'),
       value: formatCentsToReadableValue({ value: totalBorrowCents }),
     },
   ];
 
   if (displayTotalVaultStake) {
     cells.push({
-      label: t('account.marketBreakdown.cellGroup.totalVaultStake'),
+      label: t('account.summary.cellGroup.totalVaultStake'),
       value: formatCentsToReadableValue({ value: totalVaultStakeCents }),
     });
   }
 
   return (
-    <Paper css={styles.container} className={className} data-testid={TEST_IDS.container}>
-      <CellGroup
-        smallValues={displayAccountHealth}
-        cells={cells}
-        css={styles.cellGroup}
-        data-testid={TEST_IDS.stats}
-      />
+    <Section className={className} title={t('account.summary.title')}>
+      <Paper css={styles.container} data-testid={TEST_IDS.container}>
+        <CellGroup
+          smallValues={displayAccountHealth}
+          cells={cells}
+          css={styles.cellGroup}
+          data-testid={TEST_IDS.stats}
+        />
 
-      {displayAccountHealth && (
-        <div css={styles.accountHealth} data-testid={TEST_IDS.accountHealth}>
-          <BorrowLimitUsedAccountHealth
-            variant="borrowLimitUsed"
-            borrowBalanceCents={totalBorrowCents.toNumber()}
-            borrowLimitCents={borrowLimitCents.toNumber()}
-            safeBorrowLimitPercentage={SAFE_BORROW_LIMIT_PERCENTAGE}
-            css={styles.accountHealthProgressBar}
-          />
+        {displayAccountHealth && (
+          <div css={styles.accountHealth} data-testid={TEST_IDS.accountHealth}>
+            <BorrowLimitUsedAccountHealth
+              variant="borrowLimitUsed"
+              borrowBalanceCents={totalBorrowCents.toNumber()}
+              borrowLimitCents={borrowLimitCents.toNumber()}
+              safeBorrowLimitPercentage={SAFE_BORROW_LIMIT_PERCENTAGE}
+              css={styles.accountHealthProgressBar}
+            />
 
-          <div css={styles.accountHealthFooter}>
-            <Icon name="shield" css={styles.shieldIcon} />
+            <div css={styles.accountHealthFooter}>
+              <Icon name="shield" css={styles.shieldIcon} />
 
-            <Typography component="span" variant="small2" css={styles.inlineLabel}>
-              {t('myAccount.safeLimit')}
-            </Typography>
+              <Typography component="span" variant="small2" css={styles.inlineLabel}>
+                {t('myAccount.safeLimit')}
+              </Typography>
 
-            <Typography
-              component="span"
-              variant="small1"
-              color="text.primary"
-              css={styles.safeLimit}
-            >
-              {readableSafeBorrowLimit}
-            </Typography>
+              <Typography
+                component="span"
+                variant="small1"
+                color="text.primary"
+                css={styles.safeLimit}
+              >
+                {readableSafeBorrowLimit}
+              </Typography>
 
-            <Tooltip
-              css={styles.tooltip}
-              title={t('myAccount.safeLimitTooltip', { safeBorrowLimitPercentage })}
-            >
-              <Icon css={styles.infoIcon} name="info" />
-            </Tooltip>
+              <Tooltip
+                css={styles.tooltip}
+                title={t('myAccount.safeLimitTooltip', { safeBorrowLimitPercentage })}
+              >
+                <Icon css={styles.infoIcon} name="info" />
+              </Tooltip>
+            </div>
           </div>
-        </div>
-      )}
-    </Paper>
+        )}
+      </Paper>
+    </Section>
   );
 };
